+++ src/featbin/compute-spectrogram-feats.cc
@@ -2,6 +2,8 @@
 
 // Copyright 2009-2011  Microsoft Corporation
 
+// See ../../COPYING for clarification regarding multiple authors
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -138,7 +140,7 @@ int main(int argc, char *argv[]) {
         HtkHeader header = {
           features.NumRows(),
           frame_shift,
-          sizeof(float)*features.NumCols(),
+          static_cast<int16>(sizeof(float)*features.NumCols()),
           007 | 020000
         };
         p.second = header;
